Resolutions: take test-merge's mixer comments and #if U64 guard in c64.cc;
keep u64_config.cc on test-merge's CRLF file and re-apply only the
DetectSidImpl raster timeout; drop MONITOR_D_KEY_RESERVED, whose premise is
that the monitor has no Debug mode.

+++ software/io/iec/cbmdos_parser.cc
@@ -223,7 +223,7 @@ int parse_open(const char *buf, open_t& fn)
             if (i == 1) {
                 fn.filetype = e_rel;
                 if (modifiers[2]) {
-                    fn.record_size = modifiers[2][0];
+                    fn.record_size = (uint8_t)modifiers[2][0];
                 } // if not set, it will be zero
                 i++;
             } else {
